remove #define EXTERN trick in cgraph
This was designed to allow one declaration of Ag_G_global serve as both
declaration and definition. However, this obscures the actual definition of this
variable from any code browser unable to evaluate macro expansion (e.g. Cscope).
There's no real need for this trick, as we can just define the variable as
normal.