use Bison's api.prefix instead of name mangling in expr
This is the equivalent of cc2cb52756f545d6863d6fd801fa23762a81704c. My original
intent was to split the changes that remove the build-time symbol mangling (the
equivalent of 9240da12ab7857975aa34279520f1f6aad113d48) into a separate commit.
However, the expression parser does a lot of messy tricks to collide with and
suppress Bison-generated macros. It was simpler to undo all of this in one go
then try to thread this convoluted needle. Related to #1806.