Remove obsolete code within #ifdef UNUSED in lib/dotgen/aspect.c (part 1)
This is the first part of code that was obsoleted directly after the
inital commit to "Really integrate and turn on aspect ratio code in
dot".

It's a separate commit from the second part in order to facilitate
review since git shows a very strange diff otherwise.

Some comments from the removed code:

 * This code is removed. It considers dummy nodes in layer width,
 * which does not give good results in experiments.

 * The following code corrects w by adding dummy node spacing.
 * It's no longer used

 * The following commented codes are no longer used
 * Originally used in the cocktail tool.

 * The following function is no longer used.
 * Originally used for FFDH packing heuristic
 * FFDH procedure

 * packing by creating new layers. Must be commented out if packing by
 * demotion is used

 * The following code corrects w by adding dummy node spacing.
 * It's no longer used

 * The following code is for deletion of long virtual edges.
 * It's no longer used.