Remove obsolete code within #ifdef UNUSED in lib/dotgen/aspect.c (part 2)
This is the second part of code that was obsoleted directly after the
inital commit to "Really integrate and turn on aspect ratio code in
dot".

It's a separate commit from the first part in order to facilitate
review since git shows a very strange diff otherwise.

Some comments from the removed code:

 * balanceLayers:
 * The following is the layer balancing heuristic.
 * Balance the widths of the layers as much as possible.
 * It's no longer used.

 * applyPacking:
 * The following is the initial packing heuristic
 * It's no longer used.

 * applyPacking4:
 * The following is the packing heuristic for wide layout.
 * It's used with Nikolov-Healy healy heuristic.

 * rank2:
 * Initial codes for ranking (Nikolov-Healy).
 * It's no longer used.

 * rank4:
 * This function is calls the NikolovHealy function
 * for ranking in the Nikolov-Healy approach.

 * applyExpansion:
 * Heuristic for expanding very narrow graphs by edge reversal.
 * Needs improvement.