remove use of _more_long_int and _more_void_int
These macros are created by the iffe interpreter, but we do not need to perform
this kind of compile time testing. Modern compilers are clever enough to
optimize sizeof comparisons. This change uses intptr_t and uintptr_t which are
technically not available until C99, but they are used elsewhere in the codebase
already. Related to #1422 and #1521.