Do not fail if ps2pdf is present and the .pdf files exist
If ps2pdf is not present on the system, the build
fails even if the .pdf files exist. The reason is that the .ps
files are created on the fly, and they are more recent than the
distributed .pdf. Makefile then runs "false <something>", which
obviously fails.

Setting PS2PDF to "true" is not a solution because we do not
want to ignore when the .pdf files are really absent.

This patch makes running $(PS2PDF) conditional, but still
imposes the .pdf files to be generated when running "make dist".