remove explicit output filename in PDF Makefile rules
For all of these rules, the output filename ps2pdf infers is the same as the
explicit one being removed, so this change is a no-op. The intent here is to
make these invocations compatible with either ps2pdf or pstopdf filling the role
of @PS2PDF@. The usage for these two differs slightly:

  ps2pdf infile [outfile]
  pstopdf infile [-o outfile]

pstopdf is not currently supported -- there is an attempt at this in
configure.ac, but it is faulty -- but we plan to support it in an upcoming
commit. Related to #1763.