fix latent invalid pointer reference
At the point at which this code appears, trnum can be negative. This is latent
and does not cause a segfault because the compiler simply calculates a memory
address to store into t, without doing any actual dereferences. However,
debugging #56 revealed this invalid reference, which is undefined behavior with
respect to the C standard.