stop shipping strcasecmp.h
The header strcasecmp.h is a portability shim that was introduced in
cadb72d6157603338a990be4e27a599c7b96b214. However, no installed Graphviz headers
#include this header, so we do not need to install it. It is cleaner not to
clutter the shipped Graphviz include directory with an internal shim like this.

Note that there is no changelog entry for this change and this is not an API
break because there has not yet been a release since this header was added.