fix: detect absolute paths correctly in GVPR on Windows
This code was incorrectly depending on WIN32_DLL which was not defined when the
source file was built. Rather than addressing this directly, we make gvpr comply
more accurately with its man page:

  If progfile contains a slash character, the name is taken as the pathname of
  the file.

We now look for a path separator, backslash on Windows and forward slash on
anything else. Fixes #1780.