remove dead code for suffix handling of parsed numbers
This code attempted to handle a numeric suffix indicating a multiplier (e.g. 'k'
for 1024). However, the preceding logic that constructs a temporary buffer for
it to parse from does not anticipate suffixes and copy them to the buffer. The
(inactive) suffix code is considered only useful for describing file sizes, so
not relevant to enable for Graphviz. Closes #1933.