unify some array allocation across Windows and non-Windows
Just because GCC/Clang support variable length arrays, does not mean they are a
good idea. Using heap allocated memory for this is safer. This way we also get
to maintain a single code path. This adds some new compiler warnings, but
they're in line with the issues we already know (namely, passing int values as
allocation sizes).