Change Docker context to be the repo root
See https://stackoverflow.com/a/24540011/171898 and
https://docs.docker.com/engine/reference/builder/#usage

The directory passed to docker is the context. Previously we changed
directory into `ci/$IMAGE` then passed `.` as the "context".

Passing a subdir as context prevents any inclusion of files outside the
context, which isn't what we want: we want to include `requirements.txt`
from the root.

So instead, pass the entire repo as "context", and use the "-f" flag to
set the Dockerfile used.

Should help with #1743 and !1784