remove intermediate buffer in libxdot
The function jsonString was trying to optimize processing by constructing a
string in an intermediate agxbuf before passing this buffer to the caller’s
print callback. An implicit assumption is that the callback is significantly
more expensive than outputting to the intermediate buffer. This is not true. The
only callbacks ever passed to this function are agxbprint itself and fputs,
which does efficient buffering. Both of these are at least as efficient as
agxbuf, making the trip through the intermediate buffer actually a
*de-optimization*.