Redirect doc files to graphviz.org
The canonical location for these docs is in the graphviz.gitlab.io repo,
and it's extra work to maintain them in both places. The files have
gotten a bit out of sync, and have been migrated to markdown in the
other repo.

Adding the redirects ensures that they point at the latest versions with
the latest changes.