swap Strcmp calls for strcmp
Strcmp optimized strcmp by unrolling the first iteration of a byte comparison
loop. This is no longer an optimization. In modern compilers/libc, (1) strcmp is
generally no longer a byte-by-byte comparison loop and (2) the compiler knows
strcmp as a built-in and can do this kind of unrolling itself. As a result, the
compiler actually has to do *more* work to reverse Strcmp back into strcmp and
thus determine it can use, e.g., SIMD optimizations.

This is essentially the next step of f5a326cb061866bb3beaa33c3e5169efe7fc134d.