Generate shapes responsively: reflow small screens
Don't use a fixed table of width 4, reflow to the available space.

This was a bit of a pain to get the CSS right. This stackoverflow answer
was very helpful: https://stackoverflow.com/a/14155983/171898

This is towards moving shapes.html into graphviz.gitlab.io repo. Simpler
for loops that jekyll can handle.