fix: emit Bézier curves as point-by-point spline shapes in PIC plugin
With this change, the PIC plugin is finally usable. Bézier curves were being
emitted as a sequence of space-separated numbers. Similar to the previous
situation with boxes, I’m not sure what the intent of this was. This is invalid
PIC syntax. As of this commit, the PIC plugin produces output that is accepted
by GNU PIC. Fixes #131.