remove unused HAVE_ARGZ_APPEND
It is unclear where this originated from. It was added in the original commit of
this file but, at the time, the Autotools build system was not generating this
variable. The argz functionality seems related to the (now removed) dependency
RxSpencer, but Graphviz never used this and never checked whether it was
available.