Fix for TBbalance attribute code
As described in #1339 TBbalance is an undocumented graph attribute that is
supposed to cause nodes to "rise" or "fall" to the minimum or maximum rank
possible.  The attached code snippet seems to repair the bug in this code.

Comment from Matthew Fernandez: I committed this, but I don’t understand enough
about the context to judge its correctness. The above is Steve’s description of
the situation.

Fixes #1339 and #1980.