fix: don't run pylint CI job on tags
383ec8753b8007d26fc4f6eb51c61b6d8b71d12e added a pylint task in CI. This task is
unrestricted, so Gitlab thinks it is meant to run on tags as well as on master.
Every other task is restricted to only run on master (including the dependencies
of the pylint job) so CI now fails on tags. This change limits the lint job to
not run on tags. Fixes #1978.