move macOS CI jobs to Gitlab shared runner
The macOS CI closed beta program Graphviz has been participating in is winding
down. As of 2021-03-24, we have been requested to migrate to a shared macOS
runner as the program transitions into an open beta [0]. The cut over date for
this is scheduled for 2021-04-07. This change causes our CI setup to now run
macOS jobs on this shared runner instead of the reserved macOS VM Graphviz had
during the closed beta.

Note that this change also migrates these CI jobs from macOS 10.15.6 (Catalina)
to macOS 11.2 (Big Sur). With this, the XCode version is expected to move
forwards as well, though I do not know the precise version numbers.

Related to #1727.

  [0]: https://gitlab.com/gitlab-com/macos-buildcloud-runners-beta/-/issues/3#note_536852935