fix: put HTML bit at the top of the count in reference counted strings
lib/cgraph/refstr.c implements reference-counted strings with the structure
refstr_t. Since the initial commit of cgraph,
d7767d4b5c511afa89b46798eea4e2f915930844, HTML_BIT has been (IMHO incorrectly)
set based on unsigned int instead of unsigned long. Presumably this was not an
issue because most machines were 32-bit x86 at the time, where
sizeof(unsigned int) == sizeof(unsigned long).

In the transition to x86-64 machines, this setup became awkward. The reference
count is 8 bytes wide, but HTML_BIT is set to be bit 31. I.e. the bit indicating
that a string is HTML is now in the middle of the reference count. A string with
≥ 2³¹ references, something that is possible on an x86-64 machine, now
accidentally sets the HTML bit. Note that this ecosystem transition also made
the comment about HTML_BIT incorrect, /* msbit of unsigned long */.

d1244c8001e8c681def4c0ff25a91136845c2a75 did a tree-wide replacement of unsigned
long with uint64_t, which further compounded the situation. Now the HTML bit is
in the incorrect position on both 32-bit and 64-bit x86.

The present change fixes this by moving the HTML bit into bit 63 of the refcnt
field. Fixes #1984.

It is questionable whether this even needs to be a bitfield, and not simply a
uint64_t and a bool, but I think we should leave this space optimization in
place for now.