fix: do not recognize "&;" as an XML escape sequence
xml_isentity was incorrectly recognizing "&;" as an escape sequence. Despite
vague wording in the standard, it seems fairly clear that &<name>; is only
intended to be a valid escape sequence when <name> is non-empty. Fixes #797.

Note that unfortunately due to #1868, we need to fix this bug four times in
copy-pasted versions of the same function.