add a new gvc++ library with a very basic GVContext C++ class
The GVContext class is an RAII-class that handles the lifetime of the
underlying GVC_t C data structure. It currently supports construction
with or without built-in plugins, move construction and move
assignment, but not copy construction or copy assignment since the
underlying C data structure cannot be easily copied. It also supports
retrieving a non-owning pointer to the underlying GVC_t.

The new gvc++ library is optional and can be enabled with the CMake
option "with_cxx_api".

Towards https://gitlab.com/graphviz/graphviz/-/issues/2001.