install Python 3 as a pre-requisite in Windows CI tasks
Prior to this change, Python 3.8 was used on Windows in CI via the Windows
build dependencies repository.¹ This works but has a number of shortcomings:

  1. Upgrading the version of Python used in CI requires access to a Windows
     machine; and
  2. Upgrading the version of Python used in CI requires building CPython from
     source; and
  3. Python-dependent failures are hard to root cause because they may be an
     issue with how we built CPython.²

As of this commit, Python 3 is installed via the Chocolatey³ package manager.
At time of writing, this installs Python 3.9.4. This is not pinned to a specific
version – i.e. a new Chocolatey package of Python 3 will result in Graphviz
immediately using that in CI. I believe this is actually the behavior we want
right now; pro-actively catching any issues caused by Python upgrades.

This does not change the baseline requirement for Graphviz CI/test scripts from
Python 3.6. Scripts should still be written to be backwards compatible to 3.6.

As this change introduces a new non-trivial step in the Windows CI tasks,
execution time is relevant. At time of writing, the macOS jobs dominate Graphviz
CI execution time, both before and after this change. Thus I think this is
acceptable.

  ¹ https://gitlab.com/graphviz/graphviz-build-utilities
  ² E.g. #2000, “CI Python OSErrors from Windows tasks”
  ³ https://chocolatey.org/